Move Publish Schools forms into Providers module
