+++ app/controllers/publish/providers/schools/search_controller.rb
@@ -9,18 +9,18 @@ module Publish
         before_action :authorize_provider
 
         def new
-          @school_search_form = Publish::Schools::SearchForm.new
+          @school_search_form = SearchForm.new
         end
 
         def create
           redirect_to_next_step and return if school_id.present?
 
-          @school_search_form = Schools::SearchForm.new(query:)
+          @school_search_form = SearchForm.new(query:)
 
           if @school_search_form.valid?
 
-            @school_select_form = Schools::SelectForm.new
-            @school_search = Schools::SearchService.call(query:)
+            @school_select_form = SelectForm.new
+            @school_search = Publish::Schools::SearchService.call(query:)
 
             render :results
           else
@@ -29,12 +29,12 @@ module Publish
         end
 
         def update
-          @school_select_form = Schools::SelectForm.new(school_id: school_select_params[:school_id])
+          @school_select_form = SelectForm.new(school_id: school_select_params[:school_id])
 
           if @school_select_form.valid?
             redirect_to new_publish_provider_recruitment_cycle_school_path(provider_code: provider.provider_code, school_id: @school_select_form.school_id)
           else
-            @school_search = Schools::SearchService.call(query:)
+            @school_search = Publish::Schools::SearchService.call(query:)
             render :results
           end
         end
@@ -55,15 +55,15 @@ module Publish
         end
 
         def school_search_params
-          return {} unless params.key?(:publish_schools_search_form)
+          return {} unless params.key?(:publish_providers_schools_search_form)
 
-          params.expect(publish_schools_search_form: [*Schools::SearchForm::FIELDS, :school_id])
+          params.expect(publish_providers_schools_search_form: [*SearchForm::FIELDS, :school_id])
         end
 
         def school_select_params
-          return {} unless params.key?(:publish_schools_select_form)
+          return {} unless params.key?(:publish_providers_schools_select_form)
 
-          params.expect(publish_schools_select_form: [*Schools::SelectForm::FIELDS, *Schools::SearchForm::FIELDS])
+          params.expect(publish_providers_schools_select_form: [*SelectForm::FIELDS, *SearchForm::FIELDS])
         end
 
         def search_result_title_component
