+++ app/forms/publish/providers/schools/search_form.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Publish
+  module Providers
+    module Schools
+      class SearchForm
+        include ActiveModel::Model
+
+        FIELDS = %i[
+          query
+        ].freeze
+
+        attr_accessor(*FIELDS)
+
+        validates :query, presence: true, length: { minimum: 2 }
+      end
+    end
+  end
+end
