--- spec/forms/publish/schools/select_form_spec.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-module Publish
-  module Schools
-    describe SelectForm, type: :model do
-      subject { described_class.new }
-
-      include_examples 'select form'
-    end
-  end
-end
